Add files via upload
Git is upset because of a massive file I mistakenly included in a previous commit, and I can't figure out how to fix it. Even cloning the repo isn't working. It's not a compatible repo (?) or it's populated with branches from a different repo (!?). So here are the latest files, let's hope and pray to the internent jebus.